mkaurball: generate .SRCINFO, forward to makepkg if possible
Recent makepkg supports SRCINFO generation on its own, so leverage
makepkg if it's capable. Two exceptions to this rule -- if the -e or -a
flags are passed, we have no choice but to fall back on mkaurball. In
this case, warn the user that this functionality is deprecated and will
be unavailable in future version of mkaurball^Wmakepkg.

And, now that AUR 3.5.0 has been released, SRCINFO files are read in
addition to AURINFO. Let's help speed up the deprecation of AURINFO by
generating .SRCINFO files instead of .AURINFO.