array_build: clear dest array before populating
This has to be the right thing to do. Suppose some logic like:

  for l in foo bar baz; do
    if array_build a somevalues; then
      all+=("${a[@]}")
    fi
  done

If and only if arrays 'foo' 'bar' and 'baz' are in ascending order by
length, will we get this right. Otherwise, we'll overwrite a leading
portion of the destination array 'a' and then get the actual length
wrong. Clearing 'a' before populating it in array_build resolves this.